# bootcamp-server-node/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "My User Project CRUD",
        "description": "My User Project Application API",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:3000",
    "basePath": "/",
...
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/users": {
          "post": {
            "summary": "add user",
            "description": "add user to users",
            "produces": ["application/json"],
            "parameters": [
              {
                "in": "body",
                "name": "body",
                "description": "user object",
                "required": true,
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object"
                    }
                  }
                }
              }
            ],
            "responses": {
              "201": {
                "description": "created successfully",
                "schema": {
                  "type": "array"
                }
              }
            }
          }
        },
...
          "/meeting": {
            "post": {
              "summary": "add meeting",
              "description": "add meeting to meetings",
              "produces": ["application/json"],
              "parameters": [
                {
                  "in": "body",
                  "name": "body",
                  "description": "meeting object",
                  "required": true,
                  "schema": {
                    "type": "object",
                    "properties": {
                      "meeting": {
                        "type": "object"
                      }
                    }
                  }
                }
              ],
              "responses": {
                "201": {
                  "description": "created successfully",
                  "schema": {
                    "type": "array"
                  }
                }
              }
            }
          },
...
          }
        }